Express basic material and example from Brad Traversy
Reference for youtube added in the README and sample app taken from brad traversy's github
