--- express-practice/Example/getting_started_express/api/members.js
@@ -1,22 +1,0 @@
-const members = [
-    {
-        id: 1,
-        name: 'John Doe',
-        email: 'jhon@gmail.com',
-        status: 'active'
-    },
-    {
-        id: 2,
-        name: 'Bob Williams',
-        emails: 'bob@gmail.com',
-        status: 'inactive'
-    },
-    {
-        id: 3,
-        name: 'Shannon Jackson',
-        email: 'shannon@gmail.com',
-        status: 'active'
-    }
-];
-
-module.exports = members;
