--- express-practice/Example/getting_started_express/crud_api/middleware/logger.js
@@ -1,7 +1,0 @@
-const logger = (req, res, next) => {
-    // Logging the url
-    console.log(`${req.protocol}://${req.get('host')}${req.originalUrl}`);
-    next();
-}
-
-module.exports = logger;
