--- express-practice/Example/getting_started_express/helloexpress.js
@@ -1,14 +1,0 @@
-const express = require("express");
-
-// Initialize express
-const app = express();
-const PORT = process.env.PORT || 3000;
-
-app.get('/', (req, res) => {
-  res.send('Hello World');
-});
-
-// listining on a port
-app.listen(PORT, () => {
-  console.log(`Example app listening on port ${PORT}!`)
-});
