--- express-practice/Example/getting_started_express/static_server/public/jpage.json
@@ -1,4 +1,0 @@
-{
-    "cat": 27,
-    "dog": 48
-}
