--- express-practice/Example/getting_started_express/static_server/public/page2.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="styles.css">
-    <title>P2</title>
-</head>
-<body>
-    <h1>Page 2</h1>
-</body>
-</html>
