--- express-practice/Example/getting_started_express/static_server/public/styles.css
@@ -1,4 +1,0 @@
-body {
-    background-color: black;
-    color: white;
-}
