--- express-practice/Example/getting_started_express/test1/index.js
@@ -1,15 +1,0 @@
-const express = require('express');
-const path = require('path');
-
-// Initialize express
-const app = express();
-const PORT = process.env.PORT || 3000;
-
-app.get('/', (req, res) => {
-  res.sendFile(path.join(__dirname, 'public', 'index.html'));
-});
-
-// listining on a port
-app.listen(PORT, () => {
-  console.log(`Example app listening on port ${PORT}!`)
-});
