+++ express-practice/Example/index.js
@@ -1,0 +1,12 @@
+const express = require('express');
+const app = express();
+const members = require('./Members');
+const logger = require('./middleware/logger');
+
+app.use(logger);;
+
+app.use('/api/members', require('./routes/api/members'));
+
+const PORT = process.env.PORT || 5000;
+
+app.listen(PORT, () => console.log(`Listening on port ${PORT}`))
