+++ express-practice/Example/middleware/logger.js
@@ -1,0 +1,12 @@
+const moment = require('moment');
+
+const logger = (req, res, next) => {
+  console.log(
+    `${req.protocol}://${req.get('host')}${
+      req.originalUrl
+    }: ${moment().format()}`
+  );
+  next();
+};
+
+module.exports = logger;
