+++ express-practice/Example/package.json
@@ -4,14 +4,15 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "node getting_started_node/dynamic_http/index",
-    "dev": "nodemon getting_started_express/crud_api/index"
+    "start": "node index",
+    "dev": "nodemon index"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "express": "^4.17.1",
     "express-generator": "^4.16.1",
+    "moment": "^2.29.3",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
